pyproject.toml
[project]
name = "rag-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jieba>=0.42.1",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "pybase62>=1.0.0",
    "pymupdf>=1.26.3",
    "pyyaml>=6.0.2",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.13.0",
    "rouge>=1.0.1",
]
